[fix]給食の読み込みロジックの変更[refactor]enumの整理
