+++ lib/model/school/school_classification.dart
@@ -1,0 +1,6 @@
+part of 'school_model.dart';
+
+enum SchoolClassification {
+  primary,    // 小学校
+  secondary,  // 中学校
+}
