+++ lib/state/splash/loading_status.dart
@@ -1,0 +1,8 @@
+part of 'splash_state.dart';
+
+enum LoadingStatus{
+  unloading,
+  reading,
+  checkingUpdate,
+  updating,
+}
