+++ lib/view_model/multi_page/common_function.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
+import 'package:hakondate/model/school/school_model.dart';
 import 'package:hakondate/model/user/user_model.dart';
+import 'package:hakondate/repository/local/schools_local_repository.dart';
 import 'package:hakondate/view_model/multi_page/user_view_model.dart';
 
 final commonFunctionProvider = StateNotifierProvider<CommonFunction, void>((ref) => CommonFunction(ref.read));
@@ -14,12 +15,13 @@ class CommonFunction extends StateNotifier<void> {
     return day1.year == day2.year && day1.month == day2.month && day1.day == day2.day;
   }
 
-  int getIdByDay(DateTime day) {
+  Future<int> getIdByDay(DateTime day) async {
     final UserModel? user = _reader(userProvider.notifier).state.currentUser;
 
     if (user == null) throw Exception('Current user does not exist');
 
-    final int schoolId = user.schoolId;
-    return day.year * 1000000 + day.month * 10000 + day.day * 100 + schoolId;
+    final SchoolModel school = await _reader(schoolsLocalRepositoryProvider).getById(user.schoolId);
+    final int parentId = school.parentId;
+    return day.year * 1000000 + day.month * 10000 + day.day * 100 + parentId;
   }
 }
